Move chunkArray to utils
